completed task

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -10,7 +10,7 @@ int main(void)
 {
 	int i;
 	int j;
-	int fibo[50];
+	unsigned long fibo[50];
 
 	fibo[0] = 1;
 	fibo[1] = 2;
